refactor(images): rename uploadImages to cloudinaryImages and add deleteImages support
